pyproject.toml
[tool.poetry]
name = "generateimage"
version = "1.0.0"
description = "Generate images of any dimension and roughly any size for testing data"
authors = ["Anton <https://github.com/Benfey>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.19.5"
piexif = "^1.1.3"
argparse = "^1.4.0"
Pillow = "^8.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
